Moves ECPrivateKey writer back into Client
