+++ rpc/src/main/scala/org/bitcoins/rpc/serializers/BitcoindJsonWriters.scala
@@ -55,10 +55,5 @@ object BitcoindJsonWriters {
       Json.toJson(o.map { case (k, v) => (keyString(k), v) })
     }
   }
-
-  implicit val importMultiAddressWrites: Writes[RpcOpts.ImportMultiAddress] =
-    Json.writes[RpcOpts.ImportMultiAddress]
-  implicit val importMultiRequestWrites: Writes[RpcOpts.ImportMultiRequest] =
-    Json.writes[RpcOpts.ImportMultiRequest]
 }
